DFAD/requirements-movinet.txt
torch>=1.2
torchvision
torchmetrics
pytorch_lightning
tensorflow
tensorflow-hub
tqdm
visdom
numpy
pillow
scikit-learn
wandb
